requirements.txt
Authlib==1.1.0
Babel==2.11.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
cryptography==38.0.4
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Moment==1.0.5
Flask-OAuth==0.12
Flask-SQLAlchemy==3.0.2
google-auth==2.15.0
google-auth-oauthlib==0.7.1
greenlet==2.0.1
httplib2==0.21.0
identify==2.5.9
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
login==0.0.6
MarkupSafe==2.1.1
nodeenv==1.7.0
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==22.0
platformdirs==2.5.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.44
toml==0.10.2
urllib3==1.26.13
URLObject==2.4.3
virtualenv==20.17.0
Werkzeug==2.2.2
WTForms==3.0.1